🔨 [services] Move function `create2` from `entrypoints` to `services`
